+++ app/api/lesson/route.ts
@@ -29,17 +29,20 @@ export async function POST(req: NextRequest) {
                     TeacherReward: body.TeacherReward
                 },
             })
-            const notification = await prisma.notification.create({
-                data: {
-                    lessonId: newLesson.id,
-                    userId: verify.id,
-                    isRead: false,
-                    type: 'NewLesson'
+            if (verify.role != 'admin') {
+                const notification = await prisma.notification.create({
+                    data: {
+                        lessonId: newLesson.id,
+                        userId: verify.id,
+                        isRead: false,
+                        type: 'create new Lesson'
+                    }
+                })
+                if (!notification) {
+                    await prisma.lesson.delete({ where: { id: newLesson.id } })
                 }
-            })
-            if (!notification) {
-                await prisma.lesson.delete({ where: { id: newLesson.id } })
             }
+
             return NextResponse.json({ lesson: newLesson, message: 'Successfully create lesson' }, { status: 201 });
         } else {
             return NextResponse.json({ message: 'fail' }, { status: 400 });
